perf(data): keep the static catalog out of the customer bundle
src/data/index.ts assembles all 448 products at module scope, so importing
anything from it — normalizePerfume, rowToInput, buildSlug — pulled roughly
250KB of product data along. Production reads the catalog from Supabase and
never touches that array, so it was pure dead weight.

Moves the normalization into src/data/normalize.ts, which imports no product
data, and leaves src/data/index.ts as the static catalog that sync-catalog
pushes. useRemotePerfumes now imports it dynamically, only in local mode.
