+++ src/components/admin/PerfumeForm.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useEffect, useRef, useState } from "react";
 import { ArrowLeft, ImageIcon, Save, X } from "lucide-react";
 import { supabase } from "../../lib/supabase";
 import { Perfume, PerfumeCategory, PerfumeCollection, PerfumeInput, COLLECTION_LABELS } from "../../types";
-import { buildSlug } from "../../data";
+import { buildSlug } from "../../data/normalize";
 
 // ─── Types ────────────────────────────────────────────────────────────────────
